refactor(auth): JWT util 폴더 위치 변경
